势力

+++ ThreeKillGame/Assets/Script/Recruit_Scripts/ChangeAndGet.cs
@@ -36,7 +36,7 @@ public class ChangeAndGet : MonoBehaviour
         string[] ids = LoadJsonFile.forcesTableDatas[UIControl.playerForceId - 1][8].Split(',');
         for (int i = 0; i < ids.Length; i++)
         {
-            heroId = int.Parse(ids[i]);  //初始武将id获取
+            heroId = (int)float.Parse(ids[i]);  //初始武将id获取
             PutCardToPos(preparation.GetChild(i), 1);   //放置到备战位一号位置，一阶
             HeroIdChangeAndSave.pos_heroId[9 + i] = heroId;
         }
